requirements.txt
g4f
oauth2client
openai
feedparser
dateparser
newspaper3k
termcolor
google-api-python-client 
google-generativeai